requirements.txt
python-dotenv
numpy
pandas
pymongo
certifi
pymongo[srv]
scikit-learn
pyaml
dill 
mlflow
dagshub
fastapi
uvicorn

## -e .